+++ src/main/java/kitchenpos/dto/request/OrderStatusUpdateRequest.java
@@ -7,6 +7,10 @@ public class OrderStatusUpdateRequest {
     @NotNull
     private final String orderStatus;
 
+    private OrderStatusUpdateRequest() {
+        this(null);
+    }
+
     public OrderStatusUpdateRequest(String orderStatus) {
         this.orderStatus = orderStatus;
     }
