+++ src/main/java/kitchenpos/dto/request/TableNumberOfGuestsUpdateRequest.java
@@ -1,8 +1,15 @@
-package kitchenpos.dto.response;
+package kitchenpos.dto.request;
+
+import javax.validation.constraints.NotNull;
 
 public class TableNumberOfGuestsUpdateRequest {
 
+    @NotNull
     private final Integer numberOfGuests;
+
+    private TableNumberOfGuestsUpdateRequest() {
+        this(null);
+    }
 
     public TableNumberOfGuestsUpdateRequest(Integer numberOfGuests) {
         this.numberOfGuests = numberOfGuests;
